fix: better error handling
